# backend-collect/doc/api.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Terminator111/COLLECT/5.0.0
info:
  description: 软工三项目接口文档
  version: 4.0.0
  title: COLLECT-API
...
paths:
  /user/register:
    post:
      tags:
        - user
      summary: 用户注册
      description:
        通过填写用户名、密码、邮箱、手机号进行新用户注册，选择用户类型
      # parameters: ?
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
  /user/login:
    post:
      tags:
        - user
      summary: 用户登录
      description: 通过用户名和密码进行登录验证
      # parameters:? 
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFormDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              # 此处schema对应ResponseVO，下面均同，务必注意！
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserViewVO'
  /user/info/{userId}:
    get:
      tags:
        - user
      summary: 根据id获取用户简略信息
      description: 返回用户用户名和类型
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/UserViewVO"
  /user/info/list:
    post:
      tags:
        - user
      summary: 根据id列表获取简略信息列表
      description: 返回用户名和类型的列表
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                description: 用户id
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserViewVO"
  /user/view/{userId}:
    get:
      tags:
        - user
      summary: 查看用户个人信息
      description: 通过user id定位用户并返回信息
      parameters:
        - name: userId
          in: path
          required: true
          description: user id
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/UserVO"
  /user/task/finished/{userId}:
    get:
      tags:
        - user
      summary: 预览已完成任务
      description: 众包工人查看个人历史完成已交报告的任务列表/发包方查看已发布的完成的任务列表（预览）
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskViewVO"
  /user/task/unfinished/{userId}:
    get:
      tags:
        - user
      summary: 预览进行中任务
      description: 众包工人查看正在执行的任务列表/发包方查看已发布的正在执行中的任务列表（预览）
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskViewVO"
  /user/task/expired/{userId}:
    get:
      tags:
        - user
      summary: 预览逾期未交报告的任务
      description: 众包工人查看个人逾期未交报告的任务
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskViewVO"
  /user/task/report/{userId}:
    get:
      tags:
        - user
      summary: 从任务中预览报告
      description: 预览任务中的报告列表。对于众包工人和发包方点进任务可看到相应的测试报告
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: taskId
          in: query
          schema:
            type: integer
          required: true
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportViewVO"
  /user/report/{userId}:
    get:
      tags:
        - user
      summary: 预览自己的所有报告
      description: 预览众包工人目前的所有报告
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportViewVO'
  /user/workerreport/{userId}:
    get:
      tags:
        - user
      summary: 众包工人获取自己的报告
      description: 已完成报告提交的众包工人可以获取自己关于某个任务的报告
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
        - name: taskId
          in: query
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/ReportViewVO'
  /user/tag/add:
    post:
      tags:
        - user
      summary: 用户添加标签
      description: 用户添加某个标签
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
          required: true
        - name: tag
          in: query
          schema:
            type: integer
            description: 传递标签对应的整型标识
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
  /user/tag/delete:
    delete:
      tags:
        - user
      summary: 用户删除标签
      description: 用户删除某个标签
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
          required: true
        - name: tag
          in: query
          schema:
            type: integer
            description: 传递标签对应的整型标识
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
  /user/worker/context/{userId}:
    get:
      tags:
        - user
      summary: 工人获取测试的环境
      description: 工人获取个人测试的环境
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/WorkerContextVO"
  /user/worker/contextUpdate/{userId}:
    post:
      tags:
        - user
      summary: 工人修改测试的环境
      description: 工人修改个人测试的环境
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerContextDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/WorkerContextVO"
  /user/worker/radar/{userId}:
    get:
      tags:
        - user
      summary: 获取工人对应的雷达图
      description: 获取工人对应的雷达图，注意最终显示的是分数，要是1-100之间的整数
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/WorkerRadarVO"
  /user/worker/wordcloud/{userId}:
    get:
      tags:
        - user
      summary: 获取工人对应的关键词
      description: 获取工人对应的关键词,千万注意调用python后要乘以100！
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkerWordCloudVO"
...
  /user/worker/activation/{userId}:
    get:
      tags:
        - user
      summary: 获取工人对应的活跃度指数
      description: 获取工人对应活跃度指数，score在1-100之间，是整数
      parameters:
        - name: userId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/WorkerActivationVO"
...
components:
  schemas:
    ResponseVO:
      type: object
      description: api返回格式
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          type: object
...
    UserVO:
      type: object
      description: 用户
      properties:
        id:
          type: integer
        type:
          type: integer
          description: 0代表发包方，1代表众包工人
        username:
          type: string
        password:
          type: string
        email:
          type: string
        phone:
          type: string
        tagGroups:
          type: array
          items:
            type: integer
    UserDTO:
      type: object
      description: 更新用户信息
      properties:
        type:
          type: integer
          description: 0代表发包方，1代表众包工人
        username:
          type: string
        password:
          type: string
        email:
          type: string
        phone:
          type: string
        ## intention1: 要不要在众包工人注册的时候让他选一选自己的标签？
        tagGroups:
          type: array
          items:
            type: integer
    UserFormDTO:
      type: object
      description: 用户登陆表单信息
      properties:
        username:
          type: string
        password:
          type: string
    UserViewVO:
      type: object
      description: 用户简略个人信息展示
      properties:
        id:
          type: integer
        type:
          type: integer
          description: 0代表发包方，1代表众包工人，2代表管理员（注册的时候前端只有0，1两个选项）
        username:
          type: string
    WorkerContextVO:
      type: object
      description: 工人测试的环境，包含四个特征
      properties:
        id:
          type: integer
          description: 工人的id
        deviceType:
          type: string
          description: 设备类型
        osInfo:
          type: string
          description: os信息
        ramSize:
          type: string
          description: 内存大小
        netEnv:
          type: string
          description: 网络环境
    WorkerContextDTO:
      type: object
      description: 工人修改测试环境表单信息
      properties:
        deviceType:
          type: string
          description: 设备类型
        osInfo:
          type: string
          description: os信息
        ramSize:
          type: string
          description: 内存大小
        netEnv:
          type: string
          description: 网络环境
    WorkerRadarVO:
      type: object
      description: 工人的雷达图信息
      properties:
        cooperateAbility:
          type: number
          description: 报告协作能力
        reviewAbility:
          type: number
          description: 报告审查能力
        creativeAbility:
          type: number
          description: 创新能力
        findBugAbility:
          type: number
          description: 寻找bug的能力
        languageAbility:
          type: number
          description: 语言表达能力
        avgAbility:
          type: number
          description: 综合能力
    WorkerWordCloudVO:
      type: object
      description: 词云VO
      properties:
        content:
          type: string
          description: 内容
        weight:
          type: integer
          description: 权重
    WorkerActivationVO:
      type: object
      description: 工人活跃度VO
      properties:
        lastSubmitTime:
          type: string
          format: date-time
        lastThreeDaySubmitNum:
          type: integer
        lastTwoWeeksSubmitNum:
          type: integer
        lastOneMonthSubmitNum:
          type: integer
        lastHalfYearSubmitNum:
          type: integer
        score:
          type: integer
...
    TaskViewVO:
      type: object
      description: 任务预览信息
      properties:
        id:
          type: integer
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        pickedWorkerNum:
          type: integer
        tagGroups:
          type: array
          items:
            type: integer
...
    ReportViewVO:
      type: object
      description: 测试报告简略信息展示
      properties:
        id:
          type: integer
        taskId:
          type: integer
          description: 可服务于众包工人直接浏览所有测试报告
        userId:
          type: integer
          description: 众包工人的id，为了让发包方看到
        title:
          type: string
        ## change4: 我在这个类型中添加了一个score属性
        score:
          type: number
          description: 报告评分
        flawNum:
          type: integer
          description: 缺陷个数
        # 要不要加一个报告状态？
    ## change2: 这里原来是FlawVO，我将它改为FlawDTO并且删除了其中的id和reportId属性
...